unk: trying to convert model inputs to bfloat16

+++ naive/sd/pipeline.py
@@ -133,7 +133,7 @@ def generate(prompt: str,
 
             # model_output is the predicted noise
             # (Batch_Size, 4, Latents_Height, Latents_Width) -> (Batch_Size, 4, Latents_Height, Latents_Width)
-            model_output = diffusion(model_input, context, time_embedding)
+            model_output = diffusion(model_input.to(torch.bfloat16), context.to(torch.bfloat16), time_embedding.to(torch.bfloat16))
 
             if do_cfg:
                 output_cond, output_uncond = model_output.chunk(2)
